refactor flourish, forbidden knowledge, greatsword, heat weaken, herb basket, and illumination
