+++ source/gear/_gearDictionary.js
@@ -39,26 +39,12 @@ for (const file of [
 	"fever-break.js",
 	"flail.js",
 	"flame-scythes.js",
-	"flourish-base.js",
-	"flourish-bouncing.js",
-	"flourish-shattering.js",
-	"flourish-staggering.js",
-	"flourish-urgent.js",
-	"forbiddenknowledge-base.js",
-	"forbiddenknowledge-enticing.js",
-	"forbiddenknowledge-soothing.js",
-	"greatsword-base.js",
-	"greatsword-chaining.js",
-	"greatsword-distracting.js",
-	"heatweaken-base.js",
-	"heatweaken-numbing.js",
-	"heatweaken-staggering.js",
-	"herbbasket-base.js",
-	"herbbasket-enticing.js",
-	"herbbasket-guarding.js",
-	"illumination-balanced.js",
-	"illumination-base.js",
-	"illumination-inspiring.js",
+	"flourish.js",
+	"forbidden-knowledge.js",
+	"greatsword.js",
+	"heat-weaken.js",
+	"herb-basket.js",
+	"illumination.js",
 	"lance-accelerating.js",
 	"lance-base.js",
 	"lance-bashing.js",
