--- source/gear/forbiddenknowledge-enticing.js
@@ -1,68 +1,0 @@
-const { GearTemplate, CombatantReference } = require('../classes');
-const { ESSENCE_MATCH_STAGGER_ALLY, SAFE_DELIMITER } = require('../constants');
-const { getPetMove } = require('../pets/_petDictionary');
-const { changeStagger } = require('../util/combatantUtil');
-
-module.exports = new GearTemplate("Enticing Forbidden Knowledge",
-	[
-		["use", "Grant an ally @{levelUps} extra level up after combat and entice their pet to use its first move"],
-		["critical", "Reduce your target's cooldowns by @{cooldownReduction}"]
-	],
-	"Pact",
-	"Light"
-).setCost(350)
-	.setEffect(([target], user, adventure) => {
-		const { essence, pactCost, scalings: { levelUps, cooldownReduction } } = module.exports;
-		if (user.team === "delver") {
-			if (adventure.room.morale < pactCost[0]) {
-				return ["...but the party didn't have enough morale to pull it off."];
-			}
-			adventure.room.morale -= pactCost[0];
-			adventure.room.addResource(`levelsGained${SAFE_DELIMITER}${adventure.getCombatantIndex(target)}`, "levelsGained", "loot", levelUps);
-		}
-		const resultLines = [`${target.name} gains a level's worth of cursed knowledge. The party's morale falls.`];
-		if (user.essence === essence) {
-			changeStagger([target], user, ESSENCE_MATCH_STAGGER_ALLY);
-		}
-		if (user.crit) {
-			let didCooldown = false;
-			target.gear?.forEach(gear => {
-				if (gear.cooldown > 1) {
-					didCooldown = true;
-					gear.cooldown -= cooldownReduction;
-				}
-			})
-			if (didCooldown) {
-				resultLines.push(`${target.name}'s cooldowns were hastened.`);
-			}
-		}
-		const ownerIndex = adventure.getCombatantIndex(target);
-		const owner = target.team === "delver" ? target : adventure.getCombatant({ team: "delver", index: ownerIndex });
-		if (owner.pet?.type) {
-			const petMoveTemplate = getPetMove(owner.pet, 0);
-			const petRNs = { delverIndex: ownerIndex, moveIndex: 0, targetReferences: [], extras: [] };
-			petMoveTemplate.rnConfig.forEach(rnType => {
-				switch (rnType) {
-					case "enemyIndex":
-						const livingEnemyIndices = [];
-						for (let i = 0; i < adventure.room.enemies.length; i++) {
-							if (adventure.room.enemies[i].hp > 0) {
-								livingEnemyIndices.push(i);
-							}
-						}
-						petRNs.targetReferences.push(new CombatantReference(owner.team === "delver" ? "enemy" : "delver", livingEnemyIndices[adventure.generateRandomNumber(livingEnemyIndices.length, "battle")]));
-						break;
-					default:
-						petRNs.extras.push(adventure.generateRandomNumber(rnType, "battle"));
-				}
-			})
-			resultLines.push(`${target.name}'s ${owner.pet.type} uses ${petMoveTemplate.name}`, ...petMoveTemplate.effect(petMoveTemplate.selector(owner, petRNs).map(reference => adventure.getCombatant(reference)), owner, adventure, { petRNs }));
-		}
-		return resultLines;
-	}, { type: "single", team: "ally" })
-	.setSidegrades("Soothing Forbidden Knowledge")
-	.setPactCost([2, "Consume @{pactCost} morale"])
-	.setScalings({
-		levelUps: 1,
-		cooldownReduction: 1
-	});
