--- source/gear/greatsword-chaining.js
@@ -1,31 +1,0 @@
-const { GearTemplate } = require('../classes');
-const { SAFE_DELIMITER, ESSENCE_MATCH_STAGGER_FOE } = require('../constants');
-const { changeStagger, dealDamage } = require('../util/combatantUtil');
-const { damageScalingGenerator } = require('./shared/scalings');
-
-module.exports = new GearTemplate("Chaining Greatsword",
-	[
-		["use", "Deal <@{damage}> @{essence} damage to a foe and their adjacent allies"],
-		["critical", "Damage x @{critBonus}"]
-	],
-	"Offense",
-	"Wind"
-).setCost(350)
-	.setEffect((targets, user, adventure) => {
-		const { essence, scalings: { damage, critBonus } } = module.exports;
-		let pendingDamage = damage.calculate(user);
-		if (user.crit) {
-			pendingDamage *= critBonus;
-		}
-		const { resultLines, survivors } = dealDamage(targets, user, pendingDamage, false, essence, adventure);
-		if (user.essence === essence) {
-			changeStagger(survivors, user, ESSENCE_MATCH_STAGGER_FOE);
-		}
-		return resultLines;
-	}, { type: `blast${SAFE_DELIMITER}1`, team: "foe" })
-	.setSidegrades("Distracting Greatsword")
-	.setCooldown(1)
-	.setScalings({
-		damage: damageScalingGenerator(40),
-		critBonus: 2
-	});
